+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.2.0
+
+* Move mapping serializer to own package (``coosos/jms-serializer-bidirectional-relation``)
+  * Fix array keys who an auto reset
+  * Necessary to add the annotation ``Coosos\BidirectionalRelation\Annotations\SerializerBidirectionalRelation`` 
+    to your root entity which will be serialized
+* Fix array link entity to original
+* DetachEntity (based on doctrine 2.7) class for recursive detach directly in properties of UnitOfWork
+* Reorganize / Optimize code
+
 # 0.1.4
 
 * Update test